Rename columnar/ directory to array/
