+++ cpp/src/arrow/array/builder_adaptive.h
@@ -19,7 +19,7 @@
 
 #include <memory>
 
-#include "arrow/columnar/builder_base.h"
+#include "arrow/array/builder_base.h"
 
 namespace arrow {
