+++ cpp/src/arrow/array/builder_base.cc
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-#include "arrow/columnar/builder_base.h"
+#include "arrow/array/builder_base.h"
 
 #include <algorithm>
 #include <cstddef>
