+++ cpp/src/arrow/array/builder_dict.h
@@ -19,8 +19,8 @@
 
 #include <memory>
 
-#include "arrow/columnar/builder_adaptive.h"  // IWYU pragma: export
-#include "arrow/columnar/builder_base.h"      // IWYU pragma: export
+#include "arrow/array/builder_adaptive.h"  // IWYU pragma: export
+#include "arrow/array/builder_base.h"      // IWYU pragma: export
 
 namespace arrow {
