+++ cpp/src/arrow/array/builder_nested.h
@@ -20,7 +20,7 @@
 #include <memory>
 #include <vector>
 
-#include "arrow/columnar/builder_base.h"
+#include "arrow/array/builder_base.h"
 
 namespace arrow {
