+++ cpp/src/arrow/array/builder_primitive.cc
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-#include "arrow/columnar/builder_primitive.h"
+#include "arrow/array/builder_primitive.h"
 
 #include <algorithm>
 #include <cstddef>
